requirements.txt
bibtexparser==1.2.0
Django==3.1.7
django-cors-headers==3.5.0
django-heroku==0.0.0
djangorestframework==3.12.4
docx==0.2.4
docxcompose==1.3.3
docxtpl==0.14.2
gunicorn==20.1.0
pandas==1.3.3
Pillow==8.3.2
psycopg2-binary==2.9.2
PyMySQL==1.0.2
python-docx==0.8.11
requests==2.25.1
sqlparse==0.4.2
weasyprint==53.3
whitenoise==5.3.0

